Moved integrity checking drivers into separate package.
